web/requirements.txt
aiofiles==0.5.0
fastapi==0.59.0
python-dotenv==0.14.0
starlette==0.13.4
uvicorn==0.11.5
wordcookies==1.0.0
